pyproject.toml
[project]
name = "sent"
version = "0.1.0"
description = "Sentiment score"
authors = [
    {name = "KV"}
]
license = {text = "NA"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "classifier (>=2.0,<3.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "fastapi (>=0.109.0,<1.0.0)",
    "uvicorn (>=0.27.0,<1.0.0)",
    "httpx (>=0.26.0,<1.0.0)",
    "grpcio (>=1.62.0,<2.0.0)",
    "grpcio-tools (>=1.62.0,<2.0.0)"
]

[tool.poetry]
packages = [
    { include = "sent" },
    { include = "protos" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

package-mode = false
